apps/api/requirements.txt
fastapi
uvicorn[standard]
httpx
sqlalchemy
asyncpg
redis
pydantic-settings
python-jose[cryptography]
